hw11-ATM - вынесла номиналы купюр в отдельный enum-класс. Но почему-то в enum нельзя создать конструктор, принимающий short, поэтому везде в программе изменила переменные или аргументы, отвечающие за номинал на Integer. Такеж сделала описание класса ATM.

+++ hw11-ATM/src/main/java/ru/otus/chistova/ATM/CalcSum.java
@@ -4,19 +4,19 @@ import java.util.HashMap;
 import java.util.Map;
 
 public class CalcSum implements Operation  {
-    HashMap<Short, Integer> banknotes;
+    HashMap<Integer, Integer> banknotes;
 
-    public CalcSum(HashMap<Short, Integer> banknotes) {
+    public CalcSum(HashMap<Integer, Integer> banknotes) {
         this.banknotes = banknotes;
     }
 
     @Override
     public long doOperation() {
         long sum = 0;
-        Short key;
+        Integer key;
         Integer value;
         for (Map.Entry entry: banknotes.entrySet()) {
-            key = (Short) entry.getKey();
+            key = (Integer) entry.getKey();
             value = (Integer) entry.getValue();
             sum+= key*value;
         }
